tweak snapshot definition format
This format, similarly to the changes made to the parameter format, make
the JSON easier to parse in non-duck-typed languages like Go, since the
format of the value can be predicted from the key.

A separate issue I have realized, but will defer from handling for now,
is that there is not currently a way for users of the Go library to
modify the schema (e.g. by adding new key types to parameters or
snapshots) without forking the library. In the future, we should switch
to using something like https://github.com/buger/jsonparser and allowing
callbacks to be registered at runtime to handle parsing specific
objects.

For the time being, the extra effort does not make sense, since this
project supports only our internal use cases. However if the TNX format
becomes more widely used, this will need to be addressed.